==> notebooks/4_get_all_players_stats.ipynb <==
### 4th Process:

- With all the plays, get all the ids (tops and rivals) and make a feature store.
- Retrieve data for stats=yearByYear and acumulate in years.
- The accumulation is because the more the player plays, the better he could be.

### Import

import json
import requests
import pandas as pd
import numpy as np
from concurrent.futures import ThreadPoolExecutor

pd.set_option('display.max_colwidth', 200)
pd.set_option('display.max_columns', None)
pd.set_option('future.no_silent_downcasting', True)

### Utils

# Function to Process Results from Various MLB Stats API Endpoints
# Provided by the Google Colab from MLB Hackaton

def process_endpoint_url(endpoint_url, pop_key=None):
  """
  Fetches data from a URL, parses JSON, and optionally pops a key.

  Args:
    endpoint_url: The URL to fetch data from.
    pop_key: The key to pop from the JSON data (optional, defaults to None).

  Returns:
    A pandas DataFrame containing the processed data
  """
  json_result = requests.get(endpoint_url).content

  data = json.loads(json_result)

   # if pop_key is provided, pop key and normalize nested fields
  if pop_key:
    df_result = pd.json_normalize(data.pop(pop_key), sep = '_')
  # if pop_key is not provided, normalize entire json
  else:
    df_result = pd.json_normalize(data)

  return df_result

### Get Full stats

top_players_plays = pd.read_csv('mlb_top_players_plays.csv')

batter_id = top_players_plays['batter_id'].unique().tolist()
pitcher_id = top_players_plays['pitcher_id'].unique().tolist()
total_id = list(set(batter_id + pitcher_id))

# Full stats list

# Normal data
# ['airOuts', 'atBats', 'balks', 'baseOnBalls', 'battersFaced', 'blownSaves', 'catchersInterference', 'caughtStealing',
# 'completeGames', 'doubles', 'earnedRuns', 'gamesFinished', 'gamesPitched', 'gamesPlayed', 'gamesStarted', 'groundIntoDoublePlay',
# 'groundOuts', 'hitBatsmen', 'hitByPitch', 'hits', 'holds', 'homeRuns', 'inheritedRunners', 'inheritedRunnersScored', 'inningsPitched',
# 'intentionalWalks', 'leftOnBase', 'losses', 'numberOfPitches', 'outs', 'pickoffs', 'plateAppearances', 'rbi', 'runs', 'sacBunts',
# 'sacFlies', 'saveOpportunities', 'saves', 'shutouts', 'stolenBases', 'strikeOuts', 'strikes', 'totalBases', 'triples', 'wildPitches', 'wins']

# Calculated data
# 'atBatsPerHomeRun' = 'atBats' / 'homeRuns'
# 'avg' = 'hits' / 'atBats'
# 'babip' = ('hits' - 'homeRuns') / ('atBats' - 'strikeOuts' - 'homeRuns')
# 'era' = 'earnedRuns' * 9 / 'inningsPitched'
# 'groundOutsToAirouts' = 'groundOuts' / 'airOuts'
# 'hitsPer9Inn' = 'hits' * 9 / 'inningsPitched'
# 'homeRunsPer9' = 'homeRuns' * 9 / 'inningsPitched'
# 'obp' = ('hits' + 'baseOnBalls' + 'hitByPitch') / ('atBats' + 'baseOnBalls' + 'hitByPitch')
# 'ops' = 'obp' + 'slg'
# 'pitchesPerInning' = 'numberOfPitches' / 'inningsPitched'
# 'runsScoredPer9' = 'runs' * 9 / 'inningsPitched'
# 'singles' = 'hits' - 'doubles' - 'triples' - 'homeRuns'
# 'slg' = ('singles' + 2 * 'doubles' + 3 * 'triples' + 4 * 'homeRuns') / 'atBats'
# 'stolenBasePercentage' = 'stolenBases' / ('stolenBases' + 'caughtStealing')
# 'strikePercentage' = 'strikes' / 'numberOfPitches'
# 'strikeoutWalkRatio' = 'strikeOuts' / 'baseOnBalls'
# 'strikeoutsPer9Inn' = 'strikeOuts' * 9 / 'inningsPitched'
# 'walksPer9Inn' = 'baseOnBalls' * 9 / 'inningsPitched'
# 'whip' = ('baseOnBalls' + 'hits') / 'inningsPitched'
# 'winPercentage' = 'wins' / ('wins' + 'losses')

base_stats = [
    'airOuts', 'atBats', 'balks', 'baseOnBalls', 'battersFaced',
    'blownSaves', 'catchersInterference', 'caughtStealing',
    'completeGames', 'doubles', 'earnedRuns', 'gamesFinished',
    'gamesPitched', 'gamesPlayed', 'gamesStarted', 'groundIntoDoublePlay',
    'groundOuts', 'hitBatsmen', 'hitByPitch', 'hits', 'holds', 'homeRuns',
    'inheritedRunners', 'inheritedRunnersScored', 'inningsPitched',
    'intentionalWalks', 'leftOnBase', 'losses', 'numberOfPitches', 'outs',
    'pickoffs', 'plateAppearances', 'rbi', 'runs', 'sacBunts', 'sacFlies',
    'saveOpportunities', 'saves', 'shutouts', 'stolenBases', 'strikeOuts',
    'strikes', 'totalBases', 'triples', 'wildPitches', 'wins'
    ]

base_season_stats = ['season', 'player_id'] + base_stats

def fetch_player_stats(player_id: int):

    player_url = f'https://statsapi.mlb.com/api/v1/people/{player_id}/stats?stats=yearByYear'
    player_data = process_endpoint_url(player_url, "stats")
    player_stats_df = pd.DataFrame(columns=base_season_stats)

    if not player_data.empty:
        player_data = player_data.loc[0, 'splits']
        for p in player_data:
            p_dict = {
                **{'season': p['season'], 'player_id': p['player']['id']},
                **p['stat']
            }
            p_dict_filtered = {
                k: v
                for k, v
                in p_dict.items()
                if k in base_season_stats
            }
            p_data = pd.DataFrame(p_dict_filtered, index=[0])
            player_stats_df = pd.concat([player_stats_df, p_data], ignore_index=True)

        player_stats_df = (
            player_stats_df
            .astype({
                'season': int,
                'inningsPitched': float,
            })
            .sort_values('season')
        )

        player_stats_df.iloc[:, 2:] = player_stats_df.iloc[:, 2:].cumsum()

        player_stats_df = (
            player_stats_df
            .assign(
                singles = player_stats_df['hits'] - player_stats_df['doubles'] - player_stats_df['triples'] - player_stats_df['homeRuns'],
            )
        )

        player_stats_df = (
            player_stats_df
            .assign(
                atBatsPerHomeRun = player_stats_df['atBats'] / player_stats_df['homeRuns'].replace(0, np.nan),
                avg = player_stats_df['hits'] / player_stats_df['atBats'].replace(0, np.nan),
                babip = (player_stats_df['hits'] - player_stats_df['homeRuns']) / (player_stats_df['atBats'] - player_stats_df['strikeOuts'] - player_stats_df['homeRuns']).replace(0, np.nan),
                era = player_stats_df['earnedRuns'] * 9 / player_stats_df['inningsPitched'].replace(0, np.nan),
                groundOutsToAirouts = player_stats_df['groundOuts'] / player_stats_df['airOuts'].replace(0, np.nan),
                hitsPer9Inn = player_stats_df['hits'] * 9 / player_stats_df['inningsPitched'].replace(0, np.nan),
                homeRunsPer9 = player_stats_df['homeRuns'] * 9 / player_stats_df['inningsPitched'].replace(0, np.nan),
                obp = (player_stats_df['hits'] + player_stats_df['baseOnBalls'] + player_stats_df['hitByPitch']) / (player_stats_df['atBats'] + player_stats_df['baseOnBalls'] + player_stats_df['hitByPitch']).replace(0, np.nan),
                pitchesPerInning = player_stats_df['numberOfPitches'] / player_stats_df['inningsPitched'].replace(0, np.nan),
                runsScoredPer9 = player_stats_df['runs'] * 9 / player_stats_df['inningsPitched'].replace(0, np.nan),
                slg = (player_stats_df['singles'] + 2 * player_stats_df['doubles'] + 3 * player_stats_df['triples'] + 4 * player_stats_df['homeRuns']) / player_stats_df['atBats'].replace(0, np.nan),
                stolenBasePercentage = player_stats_df['stolenBases'] / (player_stats_df['stolenBases'] + player_stats_df['caughtStealing']).replace(0, np.nan),
                strikePercentage = player_stats_df['strikes'] / player_stats_df['numberOfPitches'].replace(0, np.nan),
                strikeoutWalkRatio = player_stats_df['strikeOuts'] / player_stats_df['baseOnBalls'].replace(0, np.nan),
                strikeoutsPer9Inn = player_stats_df['strikeOuts'] * 9 / player_stats_df['inningsPitched'].replace(0, np.nan),
                walksPer9Inn = player_stats_df['baseOnBalls'] * 9 / player_stats_df['inningsPitched'].replace(0, np.nan),
                whip = (player_stats_df['baseOnBalls'] + player_stats_df['hits']) / player_stats_df['inningsPitched'].replace(0, np.nan),
                winPercentage = player_stats_df['wins'] / (player_stats_df['wins'] + player_stats_df['losses']).replace(0, np.nan)
            )
        )

        player_stats_df = (
            player_stats_df
            .assign(
                ops = player_stats_df['obp'] + player_stats_df['slg'],
            )
        )

    return player_stats_df

with ThreadPoolExecutor() as executor:
    player_stats_results = list(executor.map(fetch_player_stats, total_id))

all_player_stats = pd.concat(player_stats_results, ignore_index=True)

all_player_stats.to_csv('all_player_stats.csv', index=False)

### Merge data

all_player_stats.head(5)

top_players_plays = pd.read_csv('mlb_top_players_plays.csv')
top_players_id =pd.read_csv('mlb_top_players_id.csv')

(
    top_players_plays
    .head(5)
    .filter([
        'season', 'batter_id', 'pitcher_id',
        'event_type', 'vs_RHB', 'vs_LHB',
        'vs_SHB', 'vs_RHP', 'vs_LHP'
    ])
).to_dict()

If models fail, possibility to reduce the field of action with these stats

# urL = https://raw.githubusercontent.com/MajorLeagueBaseball/google-cloud-mlb-hackathon/main/datasets/mlb-statsapi-docs/MLB-StatsAPI-Spec.json

# Batting Data (64 datos)
# requests.get(mlb_stats_url).json()['components']['schemas']['BattingData']['properties'].keys()

# Selection
# Efficiency: 'average', 'onBasePercentage', 'slugging', 'onBasePlusSlugging'
# Stats: 'hits', 'doubles', triples', 'homeRuns', 'runs', 'runsBattedIn', 'extraBaseHits', 'totalBases'
# Opportunity: 'plateAppearances', 'atBats', 'babip', 'walks', 'hitByPitch', 'intentionalWalks'

# Pitching Data (97 datos)
# requests.get(mlb_stats_url).json()['components']['schemas']['PitchingData']['properties'].keys()

# Selection
# Performance: 'earnedRunAverage', 'walksHitsPerInningPitched', 'inningsPitched', 'outsPitched', 'earnedRuns'
# Efficiency: 'strikeouts', 'walks', 'pitchesThrown', 'balls', 'strikes'
# Results: 'pitchesPerInning', 'strikePercentage', 'strikeoutWalkRatio', 'hitBatsmen', 'passedBall'

# Context
# 'wins', 'losses', 'saves', 'holds'
