==> notebooks/6_Training_MLB.ipynb <==
### Imports

import numpy as np
import joblib
import pandas as pd
from sklearn.model_selection import StratifiedKFold
from sklearn.ensemble import RandomForestClassifier
from lightgbm import LGBMClassifier
from catboost import CatBoostClassifier
from sklearn.metrics import accuracy_score
from tqdm import tqdm

### Main

df = pd.read_csv("/all_plays_merged_data.csv")
X = df.drop(columns=["event_type"])
y = df["event_type"]

event_counts = df["event_type"].value_counts()
eventos_validos = event_counts[event_counts >= 2000].index
df = df[df["event_type"].isin(eventos_validos)]

kf = StratifiedKFold(n_splits=5, shuffle=True, random_state=42)

rf_model = RandomForestClassifier(
    n_estimators=100,
    max_depth=10,
    max_samples=0.8,
    n_jobs=-1,
    random_state=42
)

def cross_val_train(model, X, y, kf, model_name):
    scores = []
    for fold, (train_idx, val_idx) in enumerate(tqdm(kf.split(X, y), total=kf.get_n_splits(), desc=f"{model_name} Training")):
        X_train, X_val = X.iloc[train_idx], X.iloc[val_idx]
        y_train, y_val = y.iloc[train_idx], y.iloc[val_idx]

        model.fit(X_train, y_train)
        preds = model.predict(X_val)
        acc = accuracy_score(y_val, preds)
        scores.append(acc)
        print(f"Fold {fold+1}: Accuracy = {acc:.4f}")

    print(f"{model_name} - Accuracy Avg: {np.mean(scores):.4f}")
    return scores

Poor prediction, can be improved with different data but is enough to test a concept

rf_scores = cross_val_train(rf_model, X, y, kf, "Random Forest")

### Final model

rf_model = RandomForestClassifier(
    n_estimators=100,
    max_depth=10,
    max_samples=0.8,
    n_jobs=-1,
    random_state=42
)

rf_model.fit(X, y)

joblib.dump(rf_model, 'model.joblib')

X.columns
